pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "java-migration-tool"
version = "0.1.0"
description = "A CLI tool for migrating Java Spring apps to Spring Boot + MongoDB using LLMs"
authors = [{ name = "Mohammad Mazraeh", email = "you@example.com" }]
requires-python = ">=3.12"
readme = "README.md"
license = { text = "MIT" }

dependencies = [
    "openai>=1.0.0",
    "javalang>=0.13.0",
    "azure-identity>=1.15.0",
    "code2prompt-rs>=0.1.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "pyautogen>=0.2.0",
    "docker>=7.0.0",
    "typing-extensions>=4.9.0",

]

[project.optional-dependencies]
dev = [
    "poethepoet",
    "black",
    "ruff",
    "pyright",
    "types-docker"
]

[project.scripts]
java-migration-tool = "java_migration_tool.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/java_migration_tool"]

[tool.poe.tasks]
format = "black ."
lint = "ruff check ."
lint-fix = "ruff check . --fix"
typecheck = "pyright"
check = { shell = "poe format && poe lint && poe typecheck" }
check-fix = { shell = "poe format && poe lint-fix" }
run = "python -m java_migration_tool.cli --mode agentic cloned_repos/spring-petclinic/"

[tool.black]
line-length = 100
target-version = ['py312']
skip-string-normalization = true

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.pyright]
include = ["src"]
pythonVersion = "3.12"
typeCheckingMode = "strict"
reportMissingImports = true
reportMissingTypeStubs = false
reportPrivateUsage = false
useLibraryCodeForTypes = true
